basic calculator can now handle float values

+++ basicAddCalculator.rb
@@ -1,6 +1,6 @@
 puts ("Enter num 1 to be added")
-a = gets().chomp().to_i();
+a = gets().chomp().to_f();
 puts ("Enter num 2 to be added")
-b = gets().chomp().to_i();
+b = gets().chomp().to_f();
 sum = a+b
 puts ("sum is "+sum.to_s())
